working on features, moving tests off of seeds.rb and onto factorybot, created overly complicated tutee system, simplifying now

+++ db/migrate/20210401193404_create_initial_tables.rb
@@ -7,7 +7,7 @@ class CreateInitialTables < ActiveRecord::Migration[5.2]
       t.string :first_name
       t.string :last_name
       t.string :gender
-      t.bigint :sid
+      t.integer :sid
       t.string :major
       t.boolean :dsp
       t.boolean :transfer
